% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_barplot.R
\name{coverage_barplot}
\alias{coverage_barplot}
\title{coverage_barplot - function to plot a barplot in the "deepskyblue" color.}
\usage{
coverage_barplot(bar_heights, bar_positions, font_size = 1, width = 0.5)
}
\arguments{
\item{bar_heights, }{a vector with heights of individual bars (defining xlim for now)}

\item{bar_positions, }{locations where bars are located (defining xlim)}

\item{font_size, }{size of font (default = 1)}

\item{width, }{width of individual bars (Default well suited for coverage is 0.5)}
}
\description{
coverage_barplot - function to plot a barplot in the "deepskyblue" color.
}
