% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomescope.R
\name{estimate_Genome_4peak2}
\alias{estimate_Genome_4peak2}
\title{Wrapper function to try fitting 4 peak model with 2 forms}
\usage{
estimate_Genome_4peak2(kmer_hist_orig, x, y, k, readlength, round, foldername)
}
\description{
Wrapper function to try fitting 4 peak model with 2 forms
}
