% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomescope.R
\name{eval_model}
\alias{eval_model}
\title{Pick between the two model forms, resolves ambiguity between which is the homozygous and which is the heterozygous peak}
\usage{
eval_model(kmer_hist_orig, nls1, nls2, round, foldername)
}
\description{
Pick between the two model forms, resolves ambiguity between which is the homozygous and which is the heterozygous peak
}
