% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomescope.R
\name{genomescope}
\alias{genomescope}
\title{genomescope - GenomeScope wrapper}
\usage{
genomescope(kmer_prof, foldername, k = 21, readlength = 150)
}
\arguments{
\item{kmer_prof}{- data.frame with kmer coverages and kmer frequencies}

\item{foldername}{- name of the output}

\item{k}{- kmer size (defaut 21)}

\item{readlength}{- read length (defaut 150)}
}
\description{
genomescope - GenomeScope wrapper
}
