% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomescope.R
\docType{data}
\name{MAX_ITERATIONS}
\alias{MAX_ITERATIONS}
\title{MAX_ITERATIONS - Max rounds on NLS}
\format{
intiger
}
\usage{
MAX_ITERATIONS
}
\description{
MAX_ITERATIONS - Max rounds on NLS
}
\keyword{datasets}
