% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomescope.R
\name{min_max}
\alias{min_max}
\title{Given mean +/- stderr, report min and max value within 2 SE}
\usage{
min_max(table)
}
\arguments{
\item{table, }{a vector of two element - mean and sd}
}
\description{
Given mean +/- stderr, report min and max value within 2 SE
}
