% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_chromosome_model_X0.R
\name{nls_2peak_male_X0}
\alias{nls_2peak_male_X0}
\title{nls_2peak_male_X0 - model for estimating X chromosome size in a species with X0 sex determination}
\usage{
nls_2peak_male_X0(
  x,
  y,
  k = 21,
  estKmercov,
  estLength,
  estR,
  max_iterations = 40
)
}
\arguments{
\item{x}{- kmer coverages}

\item{y}{- kmer frequencies}

\item{k}{- kmer size (defaut 21)}

\item{estKmercov}{- starting value for k-mer coverage}

\item{estLength}{- known genome size (estimated from female)}

\item{estR}{- known heterozygosity (estimated from female)}

\item{max_iterations}{- maximum iterations that will be done by nls algorithm}
}
\description{
nls_2peak_male_X0 - model for estimating X chromosome size in a species with X0 sex determination
}
