% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_chromosome_model_XY.R
\name{nls_2peak_male_XY}
\alias{nls_2peak_male_XY}
\title{nls_2peak_male_X0 - model for estimating X chromosome size in a species with X0 sex determination}
\usage{
nls_2peak_male_XY(x, y, k, estKmercov, estLength, estR, max_iterations)
}
\arguments{
\item{x}{- kmer coverages}

\item{y}{- kmer frequencies}

\item{k}{- kmer size (defaut 21)}

\item{estKmercov}{- starting value for k-mer coverage}

\item{estLength}{- known genome size (estimated from female)}

\item{estR}{- known heterozygosity (estimated from female)}

\item{max_iterations}{- maximum iterations that will be done by nls algorithm}
}
\description{
nls_2peak_male_X0 - model for estimating X chromosome size in a species with X0 sex determination
}
