% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_tissue_model.R
\name{nlsLM_2peak_two_tissue_model}
\alias{nlsLM_2peak_two_tissue_model}
\title{Fitting two indipendent peaks to the kmer histogram, this model aims to allow for testing of evenness of spacing (expected 1:2)}
\usage{
nlsLM_2peak_two_tissue_model(x, y, kmerEst, lengthEst, hetEst = 0.6)
}
\arguments{
\item{x, }{y vectors with coverages and theuir respective numbers of kmers (frequencies)}

\item{kmerEst, }{lengthEst, hetEst parameters that serve as initial estimates of kmer coverage, genome length and proportion of kmers in the 1n peak}
}
\description{
Fitting two indipendent peaks to the kmer histogram, this model aims to allow for testing of evenness of spacing (expected 1:2)
}
