% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomescope.R
\docType{data}
\name{NUM_ROUNDS}
\alias{NUM_ROUNDS}
\title{NUM_ROUNDS - Number of rounds before giving up}
\format{
intiger
}
\usage{
NUM_ROUNDS
}
\description{
NUM_ROUNDS - Number of rounds before giving up
}
\keyword{datasets}
