% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_tissue_model.R
\name{plot_two_tissue_model}
\alias{plot_two_tissue_model}
\title{PLotting two tissue model}
\usage{
plot_two_tissue_model(model)
}
\arguments{
\item{model}{to be plotted}
}
\description{
PLotting two tissue model
}
