% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomescope.R
\docType{data}
\name{SCORE_CLOSE}
\alias{SCORE_CLOSE}
\title{SCORE_CLOSE - Overrule if two scores are within this percent (0.05 = 5 per cent) but larger difference in het}
\format{
float
}
\usage{
SCORE_CLOSE
}
\description{
SCORE_CLOSE - Overrule if two scores are within this percent (0.05 = 5 per cent) but larger difference in het
}
\keyword{datasets}
