% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomescope.R
\docType{data}
\name{SCORE_HET_FOLD_DIFFERENCE}
\alias{SCORE_HET_FOLD_DIFFERENCE}
\title{SCORE_HET_FOLD_DIFFERENCE - Overrule heterozygosity if there is a large difference in het rate}
\format{
intiger
}
\usage{
SCORE_HET_FOLD_DIFFERENCE
}
\description{
SCORE_HET_FOLD_DIFFERENCE - Overrule heterozygosity if there is a large difference in het rate
}
\keyword{datasets}
