% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomescope.R
\name{score_model}
\alias{score_model}
\title{score model by number and percent of residual errors after excluding sequencing errors}
\usage{
score_model(kmer_hist_orig, nls, round, foldername)
}
\description{
score model by number and percent of residual errors after excluding sequencing errors
}
