% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomescope.R
\docType{data}
\name{START_SHIFT}
\alias{START_SHIFT}
\title{START_SHIFT -Coverage steps to trim off between rounds}
\format{
intiger
}
\usage{
START_SHIFT
}
\description{
START_SHIFT -Coverage steps to trim off between rounds
}
\keyword{datasets}
