% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomescope.R
\docType{data}
\name{TYPICAL_ERROR}
\alias{TYPICAL_ERROR}
\title{TYPICAL_ERROR - Typical cutoff for sequencing error}
\format{
intiger
}
\usage{
TYPICAL_ERROR
}
\description{
TYPICAL_ERROR - Typical cutoff for sequencing error
}
\keyword{datasets}
