% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomescope.R
\docType{data}
\name{VERBOSE}
\alias{VERBOSE}
\title{VERBOSE - Print out VERBOSEging messages (0/1)}
\format{
intiger
}
\usage{
VERBOSE
}
\description{
VERBOSE - Print out VERBOSEging messages (0/1)
}
\keyword{datasets}
